+++ bootstrap/handlers/secret/secret.go
@@ -24,8 +24,8 @@ import (
 	"github.com/edgexfoundry/go-mod-bootstrap/bootstrap/startup"
 	"github.com/edgexfoundry/go-mod-bootstrap/config"
 	"github.com/edgexfoundry/go-mod-bootstrap/di"
-	"github.com/edgexfoundry/go-mod-bootstrap/internal/pkg/security/authtokenloader"
-	"github.com/edgexfoundry/go-mod-bootstrap/internal/pkg/security/fileioperformer"
+	"github.com/edgexfoundry/go-mod-bootstrap/security/authtokenloader"
+	"github.com/edgexfoundry/go-mod-bootstrap/security/fileioperformer"
 
 	"github.com/edgexfoundry/go-mod-secrets/pkg"
