+++ security/fileioperformer/mocks/mock_test.go
@@ -20,7 +20,7 @@ import (
 	"os"
 	"testing"
 
-	. "github.com/edgexfoundry/go-mod-bootstrap/internal/pkg/security/fileioperformer"
+	. "github.com/edgexfoundry/go-mod-bootstrap/security/fileioperformer"
 	"github.com/stretchr/testify/assert"
 )
